+++ packages/agents-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @inkeep/agents-core
 
+## 0.22.8
+
+## 0.22.7
+
+### Patch Changes
+
+- 550d251: updated inkeep pull :)
+
 ## 0.22.6
 
 ### Patch Changes